+++ packages/vendure-plugin-stripe-subscription/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.5.3 (2024-09-18)
+
+- Don't throw error while resolving `PaymentMethodQuote.stripeSubscriptionPublishableKey` if the stripe subscription publishable key is absent (#452)
+
 # 2.5.2 (2024-08-04)
 
 - Update compatibility range (#480)